requirements.txt
torch
librosa
numpy
scikit-learn
joblib
tqdm
flask
#venv\Scripts\Activate.ps1
torchvision
Pillow
